Day 16 Part 2

+++ Day16/day16.h
@@ -7,14 +7,14 @@
 typedef struct {
     size_t length;
     size_t capacity;
-    int64_t *items;
+    int8_t *items;
 } List;
 
 List list_init(char *filepath);
 void list_free(List list);
 void list_repeat(List *list, size_t repeat_count);
 
-int64_t extract_offset(List list);
+size_t extract_offset(List list);
 void flawed_frequency_transmission(List *signal, size_t phase_count);
 
 #endif // DAY16_H_
